+++ cli/src/cli_util.rs
@@ -880,7 +880,7 @@ impl WorkspaceCommandHelper {
         self.attach_revset_evaluator(expression)
     }
 
-    fn attach_revset_evaluator(
+    pub fn attach_revset_evaluator(
         &self,
         expression: Rc<RevsetExpression>,
     ) -> Result<RevsetExpressionEvaluator<'_>, CommandError> {